routes: migrate tools domain (7 routes)
Move to src/server/routes/tools.ts:
- /api/tools (GET list)
- /api/tools/:name (GET, PUT)
- /api/tools/:name/customize POST
- /api/tools/:name/override DELETE
- /api/tool-group-policies (GET, PUT :group)

Net: server.ts -201 lines (6795 → 6594). No behaviour change.
