+++ src/server/routes/dispatcher.ts
@@ -22,6 +22,7 @@ import { imageGenerationRoutes } from "./image-generation.js";
 import { modelsRoutes } from "./models.js";
 import { preferencesConfigRoutes } from "./preferences-config.js";
 import { rolesRoutes } from "./roles.js";
+import { toolsRoutes } from "./tools.js";
 
 // Registered route arrays — populated as domain migrations land. Each
 // `routes/<domain>.ts` exports a `<domain>Routes: Route[]` array. Until all
@@ -35,6 +36,7 @@ const allRoutes: ReadonlyArray<Route> = [
 	...modelsRoutes,
 	...preferencesConfigRoutes,
 	...rolesRoutes,
+	...toolsRoutes,
 ];
 
 /**
