+++ src/server/server.ts
@@ -6,7 +6,6 @@ import http from "node:http";
 import https from "node:https";
 import path from "node:path";
 
-import { fileURLToPath } from "node:url";
 import { bobbitStateDir, bobbitConfigDir, getProjectRoot } from "./bobbit-dir.js";
 import { isSetupComplete } from "./setup-status.js";
 export { isSetupComplete };
@@ -23,7 +22,7 @@ import { checkGateDependencies } from "./agent/gate-dependency-check.js";
 import { shouldCreateWorktree } from "./agent/worktree-decision.js";
 import { RoleStore } from "./agent/role-store.js";
 import { RoleManager } from "./agent/role-manager.js";
-import { ToolManager, copyDirRecursive } from "./agent/tool-manager.js";
+import { ToolManager } from "./agent/tool-manager.js";
 
 import { getPromptSections, initPromptDirs, loadPersistedPromptSections } from "./agent/system-prompt.js";
 import { recordElapsed } from "./agent/profiling.js";
@@ -2745,206 +2744,6 @@ async function handleApiRoute(
 	}
 
 	// ── Role endpoints ─────────────────────────────────────────────
-
-	// GET /api/tools — list available agent tools (with cascade origin)
-	if (url.pathname === "/api/tools" && req.method === "GET") {
-		const projectId = url.searchParams.get("projectId") || undefined;
-		const resolved = configCascade.resolveTools(projectId);
-		const tools: Array<Record<string, unknown>> = resolved.map(r => ({ ...r.item, origin: r.origin, ...(r.overrides ? { overrides: r.overrides } : {}) }));
-		// Include MCP/external tools not covered by the config cascade
-		if (toolManager) {
-			const resolvedNames = new Set(resolved.map(r => r.item.name));
-			for (const t of toolManager.getAvailableTools()) {
-				if (!resolvedNames.has(t.name)) {
-					tools.push({ ...t, origin: "mcp" });
-				}
-			}
-		}
-		json({ tools });
-		return;
-	}
-
-	// Routes with tool :name parameter
-	const toolMatch = url.pathname.match(/^\/api\/tools\/([^/]+)$/);
-	if (toolMatch) {
-		const name = decodeURIComponent(toolMatch[1]);
-
-		if (req.method === "GET") {
-			const tool = toolManager.getToolByName(name);
-			if (!tool) { json({ error: "Tool not found" }, 404); return; }
-			json(tool);
-			return;
-		}
-
-		if (req.method === "PUT") {
-			const body = await readBody(req);
-			if (!body) { json({ error: "Missing body" }, 400); return; }
-			const ok = toolManager.updateToolMetadata(name, {
-				description: body.description,
-				group: body.group,
-				docs: body.docs,
-				detail_docs: body.detail_docs,
-				grantPolicy: body.grantPolicy,
-			});
-			if (!ok) { json({ error: "Tool not found" }, 404); return; }
-			json({ ok: true });
-			return;
-		}
-	}
-
-	// Shared helper: find which group subdirectory contains a tool by scanning YAML files.
-	function findToolGroupDir(toolName: string, toolsDir: string): string | null {
-		try {
-			const entries = fs.readdirSync(toolsDir, { withFileTypes: true });
-			for (const entry of entries) {
-				if (!entry.isDirectory()) continue;
-				const groupPath = path.join(toolsDir, entry.name);
-				try {
-					const files = fs.readdirSync(groupPath);
-					for (const file of files) {
-						if (!file.endsWith(".yaml")) continue;
-						try {
-							const raw = fs.readFileSync(path.join(groupPath, file), "utf-8");
-							// Quick check without full YAML parse
-							if (raw.includes(`name: ${toolName}`) || raw.includes(`name: "${toolName}"`)) {
-								// Verify with proper field check
-								const lines = raw.split("\n");
-								for (const line of lines) {
-									const m = line.match(/^name:\s*"?([^"\n]+)"?\s*$/);
-									if (m && m[1].trim() === toolName) return entry.name;
-								}
-							}
-						} catch { /* skip unreadable */ }
-					}
-				} catch { /* skip */ }
-			}
-		} catch { /* dir doesn't exist */ }
-		return null;
-	}
-
-	// POST /api/tools/:name/customize — copy tool group to a target scope
-	const toolCustomizeMatch = url.pathname.match(/^\/api\/tools\/([^/]+)\/customize$/);
-	if (toolCustomizeMatch && req.method === "POST") {
-		const name = decodeURIComponent(toolCustomizeMatch[1]);
-		const scope = url.searchParams.get("scope") || "server";
-		const projectId = url.searchParams.get("projectId") || undefined;
-
-		// Find the tool in the cascade to get its origin
-		const resolved = configCascade.resolveTools(projectId);
-		const source = resolved.find(r => r.item.name === name);
-		if (!source) { json({ error: "Tool not found" }, 404); return; }
-
-		// Find the groupDir by scanning tool directories to locate this tool's YAML
-		const builtinToolsDir = path.join(path.dirname(fileURLToPath(import.meta.url)), "defaults", "tools");
-		const serverToolsDir = path.join(bobbitConfigDir(), "tools");
-
-		// Find groupDir from the source layer
-		let groupDir: string | null = null;
-		let sourceToolsDir: string;
-		if (source.origin === "builtin") {
-			sourceToolsDir = builtinToolsDir;
-			groupDir = findToolGroupDir(name, builtinToolsDir);
-		} else if (source.origin === "project" && projectId) {
-			const ctx = projectContextManager.getOrCreate(projectId);
-			sourceToolsDir = ctx ? path.join(ctx.configDir, "tools") : serverToolsDir;
-			groupDir = findToolGroupDir(name, sourceToolsDir);
-		} else {
-			sourceToolsDir = serverToolsDir;
-			groupDir = findToolGroupDir(name, serverToolsDir);
-		}
-		// Fallback: try all layers
-		if (!groupDir) groupDir = findToolGroupDir(name, builtinToolsDir) || findToolGroupDir(name, serverToolsDir);
-		if (!groupDir) { json({ error: "Could not determine tool group directory" }, 400); return; }
-
-		// Determine target directory
-		let targetToolsDir: string;
-		if (scope === "project" && projectId) {
-			const ctx = projectContextManager.getOrCreate(projectId);
-			if (!ctx) { json({ error: "Project not found" }, 404); return; }
-			targetToolsDir = path.join(ctx.configDir, "tools");
-		} else {
-			targetToolsDir = serverToolsDir;
-		}
-
-		// Determine the actual source dir for copying
-		let actualSourceDir = sourceToolsDir;
-		// If the source layer doesn't have this group, try builtins then server
-		if (!fs.existsSync(path.join(actualSourceDir, groupDir))) {
-			if (fs.existsSync(path.join(builtinToolsDir, groupDir))) actualSourceDir = builtinToolsDir;
-			else if (fs.existsSync(path.join(serverToolsDir, groupDir))) actualSourceDir = serverToolsDir;
-		}
-
-		const srcDir = path.join(actualSourceDir, groupDir);
-		const destDir = path.join(targetToolsDir, groupDir);
-
-		if (!fs.existsSync(srcDir)) { json({ error: "Source tool group not found" }, 404); return; }
-
-		// Copy entire group directory (recursively handles nested files)
-		copyDirRecursive(srcDir, destDir);
-
-		json({ ok: true, groupDir }, 201);
-		return;
-	}
-
-	// DELETE /api/tools/:name/override — remove tool group override at a scope
-	const toolOverrideMatch = url.pathname.match(/^\/api\/tools\/([^/]+)\/override$/);
-	if (toolOverrideMatch && req.method === "DELETE") {
-		const name = decodeURIComponent(toolOverrideMatch[1]);
-		const scope = url.searchParams.get("scope") || "server";
-		const projectId = url.searchParams.get("projectId") || undefined;
-
-		// Determine the tools directory for the target scope
-		let targetToolsDir: string;
-		if (scope === "project" && projectId) {
-			const ctx = projectContextManager.getOrCreate(projectId);
-			if (!ctx) { json({ error: "Project not found" }, 404); return; }
-			targetToolsDir = path.join(ctx.configDir, "tools");
-		} else {
-			targetToolsDir = path.join(bobbitConfigDir(), "tools");
-		}
-
-		// Find which group directory contains this tool
-		const builtinToolsDir = path.join(path.dirname(fileURLToPath(import.meta.url)), "defaults", "tools");
-
-		// Find groupDir in the target scope (the override we're deleting)
-		let groupDir = findToolGroupDir(name, targetToolsDir);
-		// If not found in target, try builtins to at least know the group name
-		if (!groupDir) groupDir = findToolGroupDir(name, builtinToolsDir);
-		if (!groupDir) { json({ error: "Could not determine tool group directory" }, 400); return; }
-
-		const dirToRemove = path.join(targetToolsDir, groupDir);
-		if (fs.existsSync(dirToRemove)) {
-			fs.rmSync(dirToRemove, { recursive: true, force: true });
-		}
-
-		json({ ok: true });
-		return;
-	}
-
-	// ── Tool group policies ──
-
-	// GET /api/tool-group-policies
-	if (url.pathname === "/api/tool-group-policies" && req.method === "GET") {
-		const projectId = url.searchParams.get("projectId") || undefined;
-		json(configCascade.resolveToolGroupPolicies(projectId));
-		return;
-	}
-
-	// PUT /api/tool-group-policies/:group
-	const groupPolicyMatch = url.pathname.match(/^\/api\/tool-group-policies\/(.+)$/);
-	if (groupPolicyMatch && req.method === "PUT") {
-		const group = decodeURIComponent(groupPolicyMatch[1]);
-		const body = await readBody(req);
-		if (!body) { json({ error: "Missing body" }, 400); return; }
-		const validPolicies = ['allow', 'ask', 'never', 'always-allow', 'ask-once', 'always-ask', 'never-ask'];
-		if (body.policy && !validPolicies.includes(body.policy)) {
-			json({ error: `Invalid policy. Must be one of: allow, ask, never` }, 400);
-			return;
-		}
-		groupPolicyStore.setGroupPolicy(group, body.policy || null);
-		json({ ok: true });
-		return;
-	}
 
 	// ── Config: default cwd ──
